+++ test/account-manager.test.ts
@@ -68,6 +68,7 @@ function setup(t: test.TestContext) {
     clientFactory: (account) => ({ accountId: account.accountId }) as never,
     bridgeFactory: (input) => ({
       handleMessage: async () => {},
+      shutdown: async () => {},
       allowSender(senderId: string) {
         input.stateStore.setPairedSenderIds([...input.stateStore.listPairedSenderIds(), senderId]);
       },